Track event indexing progress (#3075)
# Description
Our event indexing logic currently can't handle some cases well during
restarts. Let's say a contract we index doesn't emit an event for 1
month our current logic would fetch the last indexed event from that
contract and continue indexing from there. That means on restarts we
might index ~1 month (theoretically unbounded) of blocks before we can
start building new auctions (we recently started to require the indexing
logic to have seen the tip of the chain before we can start building
auctions).
Especially on fast chains like `arbitrum` this is a huge issue.

Dusts off @squadgazzz's previous
[PR](https://github.com/cowprotocol/services/pull/2965)

# Changes
* introduced a new table `last_processed_blocks` which associates a
block number for an arbitrary string
* populated table with data for `settlements`, `ethflow_refunds`, and
`onchain_orders`

## How to test
- 1 database specific test
- actually not sure how to best test this e2e. We could spawn an
autopilot, index an event, mint a ton of empty blocks, and restart the
autopilot to trigger the code we are interested in. But it's not super
clear to me what would be a good assertion that can be tested with
reasonable effort. 🤔

---------

Co-authored-by: ilya <ilya@cow.fi>