[EASY] Don't report encoding errors for merged solutions (#3084)
# Description
The majority of solutions that get discarded due to failing
encoding/simulating/scoring are actually merged solutions. These
solutions get merged by the driver and not by the solver. The driver
does it naively so it's not really surprising that it tries to merge
solutions which are incompatible (e.g. they could use the same liquidity
source).

# Changes
To avoid alerting on false positives the solver has no control over we
now only report solutions that come directly from the solver in a
non-working state (i.e. only 1 sub-solution contained).