Quote with JIT orders API (#3083)
# Description
Closes task n1 from https://github.com/cowprotocol/services/issues/3082.

Introduces a new version of the driver's `/quote` response in the API
schema in order to support a gradual migration to the quote with jit
orders making the changes backward compatible.

The driver changes are expected in the future PRs.

# Changes
> Update the driver's OpenAPI schema quote object with an enum with the
old and new versions. Trade verifier returns error in case a new version
is received.

## How to test
Existing tests. I was thinking about how to test it properly, but it
still requires e2e tests. So, it should be implemented in the next PRs.