Fixes and simplifications for RequestSharing (#3086)
## Fixed Cache Size Metric
The `RequestSharing` component was the prime suspect for a significant
memory leak due to its metric displaying that the internal cache grows
indefinitely.

After wasting waaay too much time on this it appears that the metric
simply has a bug. If somebody calls `shared_or_else()` and there is
already an **outdated** `WeakShared` item in the cache that has not been
removed by GC we'd:
1. build a new future
2. `insert()` it into the cache
3. increase the metrics for the cache size by 1

But in this specific edge case the `insert()` would actually overwrite
an existing item in the cache and not add a new one so increasing the
metrics by 1 is actually over counting the number of cached items. Over
time this accumulated more and more and made it look like the component
had a serious memory leak.

The easy fix is to just not use `inc()` and `sub()` for adjusting the
cache size metrics but rather `set(cache.len())` since that is
absolutely fool proof.


## Removed 1 `RequestSharing` instance
The `TradeEstimator` has a `RequestSharing` and contains a `TradeFinder`
which also contains a `RequestSharing` instance. This double caching is
overly complex and due to that I already suspected it 3 times that it
might be related to a memory leak. Now only the `TradeFinder` uses
`RequestSharing` which is the component that actually sends the HTTP
requests.