Always compute fresh quote when not found (#3100)
# Description
A fix on top of the https://github.com/cowprotocol/services/pull/3096
for easier reviewing.

We want to always compute the fresh quote when provided quote_id is not
found in database. This is now enabled for all orders, not just for
ETHFlow orders.

A bit more context can be found
https://github.com/cowprotocol/services/pull/3096#issuecomment-2446885623

# Changes
<!-- List of detailed changes (how the change is accomplished) -->

- [ ] Changed behavior of the function `get_or_create_quote`

## How to test
Existing tests updated.