Start indexing settlement contract as its creation as latest (#3114)
# Description
Start indexing settlement contract as its creation as latest. Starts
from index block 0 if the contract has no deployment information (e2e
test cases which deploy the contracts from scratch only).

The problem is that when we add a new network, it starts indexing from
block 0 until it finds the first settlement contract event. Once the
first event is found, it is stored in the database. By starting from the
settlement contract creation block, we start indexing in a more
efficient way.