Simulated Token Balances for Verified Quotes (#3147)
# Description

This PR is a follow up to #3125 and uses the component introduced in the
aforementioned PR for setting up a simulated token balance using state
overrides in order for quote verification to work even when the trader
does not have sufficient balance.

# Changes

The way it works is by configuring known mapping slots for the
`mapping(address => uint256) balances` in ERC20 token contract
implementations and using this to compute the slot for overriding a
separate account's token balance (the `Spardose`), which can prefund the
trader during quote simualtions. We intentionally do not override the
trader's or solver's balance in order to not interfere with the
settlement process:

> Technically a solver could have private inventory they would like to
> use for the solution instead of transfering it to the trader.

_Posted by @MartinquaXD_

Note that the type of the state override changed slightly. This is
because it was wrong to begin with. Node implementations I tested with
(Geth and Anvil) expect both the slot and the value for state overrides
to be exactly 32-bytes long (so `H256`). I guess this feature of the
state override in the `ethrpc` crate was not used in the past and
therefore no one noticed :shrug:.

- [x] Adds command line configuration for token balance overrides
- [x] Adds logic to the trades verifier to setup state overrides for
take token balances for the configured tokens
- [x] Adds logic in the settlement simulation to try and fund the trader
if they are missing balances and balance overrides are enabled
- [x] Fixes to the `StateOverride` type and serialization

### Test Plan

Added an E2E test that uses the new token balance override feature in
order to produce a verified quote for a trader with no balances. Note
that commenting out the API arguments causes the test to fail as
expected.


> This PR is the combination of two original PRs mentioned in #3125:
> * https://github.com/nlordell/cowprotocol-services/pull/1
> * https://github.com/nlordell/cowprotocol-services/pull/2.