[EASY] Improve settlement tx cancellation logging (#3165)
# Description
ATM, it is not possible to understand whether a settlement cancellation
tx succeeded. This PR adds logs with the cancellation tx hash, so it
will be possible to check the transaction on chain later. Also,
additional context is added when working with the cancellation tx
submission since, currently, it is impossible to distinguish between the
actual settlement tx submission error.