Move settlement queue to the driver (#3129)
# Description
This is a follow-up to
https://github.com/cowprotocol/services/pull/3116, which addresses a
[suggestion](https://github.com/cowprotocol/services/pull/3116#pullrequestreview-2425935969)
of moving the settlement queue to the driver. This should help avoid
increasing the block deadline since it can still be calculated starting
from the simulation block.

# Changes

- [ ] Move the settlement queue to the driver. Each solver has its own
driver endpoint and, now, its own settlement queue.
- [ ] Calculate the deadline starting from the simulation block(revert
to the previous version).
- [ ] Do not send a settle request to the solver once the deadline is
reached.
- [ ] A new driver's config to configure the max queue size.

https://github.com/cowprotocol/infrastructure/pull/2241 should be
reverted.

## How to test
Existing tests. More driver tests will be implemented once the approach
is approved.