Driver settle queue tests (#3133)
# Description
An attempt to properly test #3129. The idea is to send multiple
`/settle` requests to the same driver to check that some will be
discarded due to the settlement queue limit.

# Changes

- [ ] `anvil` with disabled auto-mining is used to accumulate `/settle`
requests in the queue.
- [ ] Introduced a new `TooManyPendingSettlements` API error to properly
identify the cause. I couldn't find a better way to ensure the logic is
correct. Also, it might be helpful for all the parties to understand why
a specific solution/settlement was discarded.
- [ ] `default_settle_queue_size` is reduced to `2`, since 1 ongoing
settlement(dequeued) + 2 more in the queue make it `3` pending in total,
which is high enough IMO.
- [ ] Removed lifetime params from the `Settle` struct, which stores the
`/settle` endpoint response data in order to fix a compilation issue.
More details:
https://github.com/cowprotocol/services/pull/3133#discussion_r1886730784

## How to test
This is the test.

(cherry picked from commit a893c3b2c47782238e72b569badf44cf4e63985c)