Fix arbitrum access list simulation (#3174)
# Description
A PR on top of https://github.com/cowprotocol/services/pull/3173

We observed that, on Arbitrum, access list estimations occasionally fail
due to solvers (the senders of simulated transactions) needing an
exceptionally high ETH balance.

The issue arises because Arbitrum’s unusually high block gas limit,
combined with an average gas price of 0.01 GWEI, results in a
requirement of approximately 12k ETH for solvers to execute these
simulations.

To address this, this PR reduces the block gas limit to a more
reasonable value, ensuring that the access list simulation logic can
function reliably without excessive ETH balance requirements.

# Changes
<!-- List of detailed changes (how the change is accomplished) -->

- [ ] Adjusted the transaction gas limit for access list estimation on
Arbitrum

## How to test
Not sure how to properly test this. It affects only Arbitrum. But we
will still be able to see the effects on the fix by observing the logs
and whether the simulation failures stopped.

<!--
## Related Issues

Fixes #
-->