only mark tokens as unsupported based on metrics for a limited time (#3205)
# Description
Currently the bad token detection based on metrics will mark tokens as
unsupported forever. This is problematic for tokens which only have
issues temporarily. For example this can happen when the most important
pool for a token gets into a weird state or when a token gets paused or
a while.

# Changes
Adjusts the logic to freeze tokens for a configurable period of time.
Once the freeze period is over we give the token another chance (even if
the stats indicate that it's currently unsupported). To not run into
issues when a token is always bad the logic was built such that 1 more
`bad` measurement is enough to freeze the token again.
That way we can safely configure a very high `min_measurements` without
having periods where a token that was flagged as bad can issues again
because we need to get a lot new measurements to mark it as unsupported
again.

Additionally the PR simplifies how the metrics based bad token detector
gets instantiated and gives each solver their completely separate
instance (how it was originally communicated because each solver may
support different tokens).

## How to test
added a unit test