Get token first trade block API (#3197)
# Description
This is a pre-requisite for improving `internal buffers at risk` alert
by eliminating false positives. The idea is to provide the token's first
trade timestamp so https://github.com/cowprotocol/tenderly-alerts/ can
then decide whether the token should be ignored. Currently, there is no
way to return the timestamp since the `order_events` table gets cleaned
up periodically. Technically, it can be used to get a timestamp, but the
result will be useless if a token was deployed more than 30 days ago and
wasn't traded for the last 30 days. Instead, the block number is
returned, so Tenderly RPC must be used to fetch the block's timestamp.
Otherwise, this would require access to an archive node from the
orderbook side.

## Changes
For some reason, this query takes around 20s on mainnet-prod for
avg-popular tokens. Unfortunately, this couldn't be reproduced locally.
I assume this is related to available resources on prod. I added indexes
that improved the query speed ~x2.5 times.

## Caching
To avoid querying the DB for the same tokens too often, I would consider
introducing caching on the NGINX side rather than in memory since we
often have multiple orderbook instances. Also, the first trade block
never changes and can be kept in the NGINX cache forever. All the errors
won't be kept in the cache. Requires an infra PR.

## How to test
The query is tested on prod and locally. This would require further
testing on prod by collecting metrics and adjusting resources.
Potentially, `work_mem`, `max_parallel_workers_per_gather`, etc.