Added `auction_id` field to `/settle` and `/reveal` driver requests. (#3212)
# Description
This is a continuation of the previous PR #3131 which added `auction_id`
field as optional on the driver `/settle` and `/reveal` requests.
Because this is API breaking change we had to wait for Solvers teams to
updated their driver version, so they will support new `auction_id`
field.
Current PR changes `auction_id` field from optional to mandatory and
starts sending it from the autopilot.

# Changes
This is mainly port of PR #3113 which introduced API break (and was
reverted with PR #3131) on to the current main branch: updated driver
and autopilot DTOs, added code to use these new fields and added
additional condition on driver side to verify if provided `solution_id`
is competing in provided `auction_id`.
OpenAPI definition is already up to date.

## How to test
Existing tests, esp. `driver_handles_solutions_based_on_id()`.