Bad token detector metrics (#3228)
# Description
Adds Prometheus metrics for driver's bad token detector strategies, so
it will be possible to show how many tokens were marked as `bad` by each
solver and strategy. It would be helpful to identify solvers that
perform as not expected and to analyze the work of each detection
strategy individually.

## How to test
Create 2 grafana panels: the one that shows detected tokens by solver
and another - by strategy.

Added `hotfix` label to deploy it asap.