[TRIVIAL] Fix API error message (#3246)
# Description
The API expects camelCase query parameters and parses them into
snake_case Rust identifiers. Therefor the `API` actually expects the
query parameter `orderUid` and not `order_uid` which should be reflected
in the error message.

# Changes
Updated error message to be correct from the perspective of the API
consumer not the backend internals.