Use foundry stable version in the playground docker image (#3255)
# Description
It was noticed by @ahhda that the test playground CI job [started to
fail
constantly](https://github.com/cowprotocol/services/actions/runs/12898722916).
Seems like one of the nightly [foundry-rs
releases](https://github.com/foundry-rs/foundry/releases) was the main
reason for this. Even though the
[stable](https://github.com/foundry-rs/foundry/releases/tag/stable)
release is marked as `latest`, according to the logs, a dev version of
`anvil` was used.

# Changes

- [ ] Switch from the `latest` to `stable` version.
- [ ] Fix autopilot driver config updated due to
https://github.com/cowprotocol/services/pull/3065

## How to test
https://github.com/cowprotocol/services/actions/runs/13018157686?pr=3254