Driver's auction preprocessing `total` metric fix (#3261)
The auction preprocessing job runs only once for each auction, which is
achieved by using a mutex where other threads wait until the running job
is completed. The `total` metric was placed incorrectly outside the
lock, accumulating a lot of useless information where the preprocessing
is already done, and other threads just read the value. This PR fixes it
by only populating the metric when the async job is really executed.