Use unbuffered transport for the Debug RPC API (#3285)
## Description

Since Alchemy is used as one of the failover providers and it doesn't
support buffered debug RPC API, it is important to use unbuffered
transport when sending `debug_traceTransaction` calls. Otherwise, this
leads to failed RPC calls.

## Changes

The easiest workaround is to use 2 separate transports for different
purposes. The current approach is not really flexible since it is not
configurable, but in order to keep it simple, this should be acceptable.